requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.7
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0

passlib~=1.7.4
bcrypt==3.2.0
auth~=0.5.3
python-jose~=3.3.0
python-multipart
fastapi-limiter
redis~=5.3.0b4
pydantic-settings~=2.7.1
alembic~=1.14.1